[Maintenance] Move order related commands to separate namespace
