+++ src/Sylius/Bundle/ApiBundle/Command/Order/ResendShipmentConfirmationEmail.php
@@ -11,16 +11,12 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\Command\Checkout;
+namespace Sylius\Bundle\ApiBundle\Command\Order;
 
 use Sylius\Bundle\ApiBundle\Attribute\ShipmentIdAware;
+use Sylius\Bundle\CoreBundle\Command\ResendShipmentConfirmationEmail as BaseResendShipmentConfirmationEmail;
 
 #[ShipmentIdAware]
-class ShipShipment
+class ResendShipmentConfirmationEmail extends BaseResendShipmentConfirmationEmail
 {
-    public function __construct(
-        public readonly mixed $shipmentId = null,
-        public readonly ?string $trackingCode = null,
-    ) {
-    }
 }
