+++ src/Sylius/Bundle/ApiBundle/CommandHandler/Order/ChangePaymentMethodHandler.php
@@ -11,10 +11,10 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\CommandHandler\Account;
+namespace Sylius\Bundle\ApiBundle\CommandHandler\Order;
 
 use Sylius\Bundle\ApiBundle\Changer\PaymentMethodChangerInterface;
-use Sylius\Bundle\ApiBundle\Command\Account\ChangePaymentMethod;
+use Sylius\Bundle\ApiBundle\Command\Order\ChangePaymentMethod;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Repository\OrderRepositoryInterface;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
