+++ src/Sylius/Bundle/ApiBundle/CommandHandler/Order/SendShipmentConfirmationEmailHandler.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
+namespace Sylius\Bundle\ApiBundle\CommandHandler\Order;
 
-use Sylius\Bundle\ApiBundle\Command\Checkout\SendShipmentConfirmationEmail;
+use Sylius\Bundle\ApiBundle\Command\Order\SendShipmentConfirmationEmail;
 use Sylius\Bundle\CoreBundle\Mailer\ShipmentEmailManagerInterface;
 use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Core\Repository\ShipmentRepositoryInterface;
