+++ src/Sylius/Bundle/ApiBundle/CommandHandler/Order/ShipShipmentHandler.php
@@ -11,11 +11,11 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
+namespace Sylius\Bundle\ApiBundle\CommandHandler\Order;
 
 use Sylius\Abstraction\StateMachine\StateMachineInterface;
-use Sylius\Bundle\ApiBundle\Command\Checkout\SendShipmentConfirmationEmail;
-use Sylius\Bundle\ApiBundle\Command\Checkout\ShipShipment;
+use Sylius\Bundle\ApiBundle\Command\Order\SendShipmentConfirmationEmail;
+use Sylius\Bundle\ApiBundle\Command\Order\ShipShipment;
 use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Core\Repository\ShipmentRepositoryInterface;
 use Sylius\Component\Shipping\ShipmentTransitions;
