+++ src/Sylius/Bundle/ApiBundle/spec/CommandHandler/Order/ShipShipmentHandlerSpec.php
@@ -11,12 +11,12 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
+namespace spec\Sylius\Bundle\ApiBundle\CommandHandler\Order;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Abstraction\StateMachine\StateMachineInterface;
-use Sylius\Bundle\ApiBundle\Command\Checkout\SendShipmentConfirmationEmail;
-use Sylius\Bundle\ApiBundle\Command\Checkout\ShipShipment;
+use Sylius\Bundle\ApiBundle\Command\Order\SendShipmentConfirmationEmail;
+use Sylius\Bundle\ApiBundle\Command\Order\ShipShipment;
 use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Core\Repository\ShipmentRepositoryInterface;
 use Sylius\Component\Shipping\ShipmentTransitions;
